requirements.txt
docling
lancedb
openai
pydantic
python-dotenv
streamlit
tiktoken